Changed tString Prefix and Suffix to Left and Right for count based actions
Prefix(int count) and Suffix as well as the Extract versions are now called Left and Right and will, unlike before, now give you what it can if you ask for more than the string has to give.  Before it gave you an empty string.  You will still get an empty string with mid if the start is < 0 or >= length.  In all cases a count of <= 0 will return an empty string.